+++ src/main/java/com/fitbum/repositorios/programa/MicrocicloRepositorio.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MicrocicloRepositorio extends JpaRepository<Microciclo,Long> {
+public interface MicrocicloRepositorio extends JpaRepository<Microciclo,Integer> {
 }
