+++ src/main/java/com/fitbum/repositorios/programa/SerieRepositorio.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SerieRepositorio extends JpaRepository<Serie,Long> {
+public interface SerieRepositorio extends JpaRepository<Serie,Integer> {
 }
